Added an arrow

+++ scripts/health.gd
@@ -1,11 +1,13 @@
 extends Node
 class_name Health
 
+@onready var animation_player: AnimationPlayer = $AnimationPlayer
+
 @export var max_health = 3
 @export var immunity_frame_duration = 1.0
 
 var last_hit_time : float
-
+var _is_dead = false
 var current_health : int
 
 signal on_death
@@ -16,8 +18,9 @@ func _ready() -> void:
 	
 func Damage(amount : int):
 	var current_time = Time.get_ticks_msec() / 1000.0
-	if current_time - last_hit_time > immunity_frame_duration:
+	if current_time - last_hit_time > immunity_frame_duration and not _is_dead:
 		current_health -= amount
+		animation_player.play("flicker")
 		last_hit_time = current_time
 		send_signal()
 	
@@ -33,6 +36,7 @@ func reset_health():
 	send_signal()
 	
 func die():
+	_is_dead = true
 	on_death.emit()
 	
 func send_signal():
